quantile calculates changed from integers to float-points, fix splited tags in grok parser

+++ src/metric/percentile_heap.h
@@ -3,7 +3,7 @@
 #include "dstructures/ht.h"
 typedef struct percentile_buffer
 {
-	int64_t *arr;
+	double *arr;
 	size_t n;
 	int64_t cur;
 	int64_t sortsize;
@@ -13,8 +13,8 @@ typedef struct percentile_buffer
 } percentile_buffer;
 
 percentile_buffer* init_percentile_buffer(int64_t *percentile, size_t n);
-void heapSort(int64_t *arr, int64_t n);
+void heapSort(double *arr, int64_t n);
 int64_t* percentile_init_3n(int64_t n1, int64_t n2, int64_t n3);
-void heap_insert(percentile_buffer *pb, int64_t key);
+void heap_insert(percentile_buffer *pb, double key);
 void calc_percentiles(void *carg, percentile_buffer *pb, void *mnode, char *custom_mname, alligator_ht *custom_labels);
 void free_percentile_buffer(percentile_buffer *pb);
